[추가] SignUp의 로직을 수정합니다.
