+++ Photory/AuthView.swift
@@ -1,0 +1,20 @@
+//
+//  AuthView.swift
+//  Photory
+//
+//  Created by Hamlit Jason on 2022/08/15.
+//
+
+import SwiftUI
+
+struct AuthView: View {
+    var body: some View {
+        SignInView()
+    }
+}
+
+struct AuthView_Previews: PreviewProvider {
+    static var previews: some View {
+        AuthView()
+    }
+}
