+++ Photory/Commons/SignHeaderView.swift
